uv_python/basic/pyproject.toml
[project]
name = "uv-python"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "fastapi>=0.127.0",
    "mysql-connector-python>=9.5.0",
    "nanoid>=2.0.0",
    "uvicorn>=0.40.0",
]
